+++ client/Packages/com.beamable.server/Editor/Usam/Editor/UsamService_Reload.cs
@@ -4,6 +4,7 @@ using Beamable.Editor.BeamCli.Commands;
 using Beamable.Editor.BeamCli.Extensions;
 using System.Linq;
 using UnityEditor;
+using UnityEngine;
 
 namespace Beamable.Server.Editor.Usam
 {
@@ -25,6 +26,18 @@ namespace Beamable.Server.Editor.Usam
 			_ctx.OnRealmChange += OnBeamEditorRealmChanged;
 
 			var command = _cli.UnityManifest();
+
+			_requiredUpgrades.Clear();
+			var checkCommand = _cli.ChecksScan(new ChecksScanArgs());
+			checkCommand.OnStreamCheckResultsForBeamoId(cb =>
+			{
+				if (latestReloadTaskId != taskId)
+					return;
+				_requiredUpgrades.Add(cb.data);
+			});
+			checkCommand.Run();
+			
+			
 			command.OnStreamShowManifestCommandOutput(cb =>
 			{
 				if (latestReloadTaskId != taskId)
@@ -32,7 +45,6 @@ namespace Beamable.Server.Editor.Usam
 				
 				hasReceivedManifestThisDomain = true;
 				latestManifest = cb.data;
-				_requiredUpgrades.Clear();
 				
 				foreach (var service in latestManifest.services)
 				{
@@ -43,11 +55,6 @@ namespace Beamable.Server.Editor.Usam
 						service.Flags |= BeamManifestEntryFlags.IS_READONLY;
 					}
 
-					var upgrades = CheckForRequiredUpgrades(service);
-					if (upgrades != null)
-					{
-						_requiredUpgrades.Add(upgrades);
-					}
 				}
 
 				foreach (var storage in latestManifest.storages)